move to legacy: stitch
