+++ weave/tests/legacy/test_stitch.py
@@ -8,7 +8,8 @@ from weave.legacy import context_state as _context
 from weave.legacy.language_features.tagging import make_tag_getter_op
 from weave.legacy.ops_domain import run_ops
 
-from ... import stitch, weave_internal
+from ... import weave_internal
+from ...legacy import stitch
 from .. import fixture_fakewandb as fwb
 from . import test_wb
