+++ weave/tests/legacy/test_wb.py
@@ -7,9 +7,8 @@ import pytest
 import wandb
 
 from weave import query_api as weave
-from weave import stitch
 from weave import weave_types as types
-from weave.legacy import artifact_fs, artifact_wandb, compile, graph, ops, uris
+from weave.legacy import artifact_fs, artifact_wandb, compile, graph, ops, stitch, uris
 from weave.legacy import ops_arrow as arrow
 from weave.legacy.language_features.tagging.tagged_value_type import TaggedValueType
 from weave.legacy.ops_arrow import ArrowWeaveListType
